D-25: the RSOC trio and the fuel cell's table path move to obsolete/
Two manufacturer tables the hydrogen chain reads are not in this repository, and as far
as its history shows never were: `hisim/inputs/fuel_cell_manufacturer_config.json` and
`hisim/inputs/rSOC_manufacturer_config.json`. `hisim/inputs/` ships
`electrolyzer_manufacturer_config.json` and `rSOC_efficiency_curve_data.json` and no
sibling of either. Every builder that opened one of the two raised `FileNotFoundError` on
its first call; the tests of those builders passed only because they fabricated a dict and
handed it in through a parameter added for the purpose.

For the RSOC trio that is total, so the trio moves. `generic_rsoc` (`RsocConfig`, `Rsoc`),
`controller_l1_rsoc` (`RsocControllerConfig`) and `controller_l2_rsoc_battery_system`
(`RsocBatteryControllerConfig` and the `RsocBatteryOperationMode` enum #687 added) go to
`obsolete/components/` with their three tests, unchanged, because none of the three
classes has any other builder: `read_config`, `from_rsoc_name` and the two `config_rsoc`
all read the absent file. `rSOC_efficiency_curve_data.json` travels with them into a new
`obsolete/inputs/`: `generic_rsoc.py` was the only module in the tree that opened it, so
with the component gone nothing in `hisim/` reads it, and a table whose only reader has
left is not an input any more.

For the fuel cell the table path moves and the PEM presets stay. `generic_fuel_cell`,
`controller_l1_fuel_cell` and `controller_l2_xtp_fuel_cell_ems` each had a `read_config`
opening the absent fuel-cell table and a factory calling it — `config_fuel_cell`,
`control_fuel_cell`, `control_fuel_cell`. All six functions are cut out and archived
verbatim in one new file, `obsolete/components/fuel_cell_manufacturer_table.py`, whose
docstring says where each came from and that its data file never existed. The three
modules, their components and their tests stay in `hisim/` and `tests/`, and the two
hand-typed defaults that do build — `FuelCellConfig.get_default_pem_fuel_cell_config` and
`FuelCellControllerConfig.get_default_fuel_cell_controller_config` — are untouched. The
two fuel-cell tests never exercised `read_config`, so neither needed a case removed.

`XTPControllerConfig` is the one that pays for this, and it is worth saying plainly: the
table path was its *only* factory, so the class now has **no default builder at all**. It
keeps its fields, its `XtpOperationMode` enum, its component and its two tests — which
build configs with explicit kwargs and still pass — and a caller must name every field
until the B7 conversion batch gives it a preset. Its class docstring now says so.

D-26 narrows as a consequence. "Raise on an unknown device" now has three readers to
normalise rather than five, and all three read the one manufacturer table that does exist:
`generic_electrolyzer_h2`, `controller_l1_electrolyzer_h2` and the PtX controller in
`controller_l2_ptx_energy_management_system`, all on
`electrolyzer_manufacturer_config.json`.

Registry and docs follow the classes. `DeviceEnergySpecs.BY_CLASS_NAME` loses its `Rsoc`
and `RsocBatteryController` rows and `tests/test_economics_extraction.py` the matching two,
because that table's contract test requires every key to name a class still living in
`hisim.components`. `docs/modules/components.rst` loses its three RSOC entries and
`docs/modules/inputs.rst` the efficiency-curve bullet. `cost_module_issues.md` item 34
listed `generic_rsoc.Rsoc` among the PRICED components with no cost-facts source; the
reversible cell is struck there with a note, leaving the fuel cell as the only half of that
bullet still open. `roadmap/` decision records are history and stay as written, except
`p4_component_sweep_requirements.md`, where D-25 is recorded as executed and the R3 rows
for the trio, the two fuel-cell configs and the XTP controller are brought into line.

Nothing is deleted and no recorded result moves: no system setup, energy system, scenario
JSON or golden reference names any of these classes — the `"FuelCell"` strings in
`dynamic_components.scenario.json` are `ComponentType.FUEL_CELL` tags on the CHP of
`advanced_fuel_cell.py`, which is a different component and is untouched.
