+++ obsolete/README.md
@@ -36,6 +36,11 @@ file may sit here for a while before disappearing from this one.
 | `components/controller_l1_electrolyzer.py` | D-29 | `L1GenericElectrolyzerController`, written for the electrolyzer above and pointed at the storage above. The follow-up decision was that it moves rather than being re-pointed at the survivor. Its test is the electrolyzer test that moved with the component; `tests/test_controller_l1_generic_electrolyzer.py` stays behind despite its name, because it builds the live `controller_l1_electrolyzer_h2.ElectrolyzerController`. |
 | `components/configuration_hydrogen.py` | D-29 | `HydrogenStorageConfig` and `AdvElectrolyzerConfig`, moved out of `hisim/components/configuration.py`, which keeps everything else — the two D-16 left behind as belonging here. Both verified to have zero references; `AdvElectrolyzerConfig` was the dead third copy of the electrolyzer configuration, disagreeing with the live one by a missing underscore (`max_power = 2_4000`). |
 | `components/generic_smart_device.py` | D-30 | Defective as well as dead: `SmartDevice.__init__` calls `build()` unconditionally, which reads `utils.HISIMPATH["utsp_reports"]` — a key `HISIMPATH` does not have (it has `utsp_results`, `utsp_example_results`, `utsp_example_reports` and `report`) — so every construction raises `KeyError: 'utsp_reports'`. Zero setups, zero call sites, and no test file, so nothing moves beside it. Even with the path fixed, the naming supplement's proposed `standard` preset would have shipped `identifier="Identifier"`, which matches no device in the LPG flexibility report, so the preset would have built an appliance with an empty profile. A D13 dead-code member the plan did not list. The KPI vocabulary keeps its `ComponentType.SMART_DEVICE` member in `loadtypes.py` and its `KpiTagEnumClass.SMART_DEVICE` tag, and the energy management system keeps the consumer branch that names the former; `SingletonDictKeyEnum.SMARTDEVICESINCLUDED` stays too — nothing in the repository ever wrote or read it, this module included, so it is D-32's business, not this move's. |
+| `components/generic_rsoc.py` | D-25 | The reversible solid-oxide cell, unbuildable since it was written. `RsocConfig.read_config` and `from_rsoc_name` — its only builders — open `hisim/inputs/rSOC_manufacturer_config.json`, which is not in this repository and, as far as its history shows, never was; both raise `FileNotFoundError`. The four tests passed only by handing the class a dict they fabricated themselves. |
+| `components/controller_l1_rsoc.py` | D-25 | `RsocControllerConfig`, the L1 controller of the cell above, with the same single builder reading the same absent file. Seven tests, all on an injected dict. |
+| `components/controller_l2_rsoc_battery_system.py` | D-25 | `RsocBatteryControllerConfig` and its `RsocBatteryOperationMode` enum. Same absent file, same total block: ten tests, all on an injected dict. Its `dataclasses_json` field aliases — seven power fields that serialize as `nom_load_soec_in_kW` while their Python names are `…_in_kw` — are the repository's only ones, and they leave with it. |
+| `inputs/rSOC_efficiency_curve_data.json` | D-25 | Travels with its only reader. `generic_rsoc.py` was the sole module that opened this efficiency curve; with the component gone, nothing in `hisim/` reads it, so it moves rather than sitting in `hisim/inputs/` as a table for a component that is no longer there. |
+| `components/fuel_cell_manufacturer_table.py` | D-25 | Not a component: the six functions of the fuel-cell manufacturer-table path, cut out of three live modules and archived verbatim. `FuelCellConfig.read_config`/`config_fuel_cell` (`generic_fuel_cell.py`), `FuelCellControllerConfig.read_config`/`control_fuel_cell` (`controller_l1_fuel_cell.py`) and `XTPControllerConfig.read_config`/`control_fuel_cell` (`controller_l2_xtp_fuel_cell_ems.py`) all opened `hisim/inputs/fuel_cell_manufacturer_config.json`, which does not exist and never did. The three modules stay in `hisim/`: the first two keep their hand-typed PEM defaults, which build. The third does not — `XTPControllerConfig` now has **no default builder at all** until the conversion batch gives it a preset, and a caller must name every field. |
 
 Two things D-16 deleted rather than moved are not in the table, because a factory is code and not a class:
 `RandomNumbersConfig.get_default_config` and `PVSystem.get_default_config`, both with zero callers. One
@@ -51,4 +56,5 @@ instead of re-pointing it, and no replacement connection to the surviving `Hydro
 controller keeps its optional `HydrogenSOC` input — a setup that wants a fuel cell to watch a hydrogen store
 wires it explicitly.
 
-`tests/` holds the tests that came with them, so their history stays next to the code they tested.
+`tests/` holds the tests that came with them, so their history stays next to the code they tested, and
+`inputs/` the data files whose only reader moved.
