+++ obsolete/components/generic_rsoc.py
