+++ obsolete/inputs/rSOC_efficiency_curve_data.json
