+++ obsolete/tests/test_controller_l1_rsoc.py
