+++ obsolete/tests/test_controller_l2_rsoc_battery_system.py
