+++ obsolete/tests/test_generic_rsoc.py
